+++ visualization/app/codeCharta/codeCharta.model.ts
@@ -25,10 +25,10 @@ export enum LayoutAlgorithm {
 }
 
 export enum SharpnessMode {
-	Standard = "Standard",
-	PixelRatioNoAA = "Pixel Ratio without Antialiasing",
-	PixelRatioFXAA = "Pixel Ratio With FXAA Antialiasing",
-	PixelRatioAA = "Pixel Ratio with Antialisaing (best)"
+	Standard = "High",
+	PixelRatioNoAA = "Low",
+	PixelRatioFXAA = "Medium",
+	PixelRatioAA = "Best"
 }
 
 export interface CCFile {
@@ -119,7 +119,7 @@ export interface DynamicSettings {
 	colorMetric: string
 	distributionMetric: string
 	edgeMetric: string
-	focusedNodePath: string
+	focusedNodePath: string[]
 	searchPattern: string
 	margin: number
 	colorRange: ColorRange