+++ visualization/app/codeCharta/state/store/dynamicSettings/focusedNodePath/focusedNodePath.selector.ts
@@ -1,0 +1,4 @@
+import { createSelector } from "../../../angular-redux/store"
+import { dynamicSettingsSelector } from "../dynamicSettings.selector"
+
+export const focusedNodePathSelector = createSelector([dynamicSettingsSelector], dynamicSettings => dynamicSettings.focusedNodePath)