+++ visualization/app/codeCharta/util/algorithm/treeMapLayout/treeMapGenerator.spec.ts
@@ -34,7 +34,7 @@ describe("treeMapGenerator", () => {
 		codeMapNode = klona(VALID_NODE_WITH_PATH)
 		metricData = klona(METRIC_DATA)
 		isDeltaState = false
-		state.dynamicSettings.focusedNodePath = ""
+		state.dynamicSettings.focusedNodePath = []
 	}
 
 	describe("create Treemap nodes", () => {